requirements.txt
numpy
pandas
psycopg2-binary
python-dotenv
